+++ examples/voice-assistant/minimal_assistant.py
@@ -1,6 +1,7 @@
 import asyncio
 
 from dotenv import load_dotenv
+from livekit import rtc
 from livekit.agents import AutoSubscribe, JobContext, WorkerOptions, cli, llm
 from livekit.agents.voice_assistant import VoiceAssistant
 from livekit.plugins import deepgram, openai, silero
@@ -27,6 +28,21 @@ async def entrypoint(ctx: JobContext):
         chat_ctx=initial_ctx,
     )
     assistant.start(ctx.room)
+
+    # listen to incoming chat messages, only required if you'd like the agent to
+    # answer incoming messages from Chat
+    chat = rtc.ChatManager(ctx.room)
+
+    async def answer_from_text(txt: str):
+        chat_ctx = assistant.chat_ctx.copy()
+        chat_ctx.append(role="user", text=txt)
+        stream = assistant.llm.chat(chat_ctx=chat_ctx)
+        await assistant.say(stream)
+
+    @chat.on("message_received")
+    def on_chat_received(msg: rtc.ChatMessage):
+        if msg.message:
+            asyncio.create_task(answer_from_text(msg.message))
 
     await asyncio.sleep(1)
     await assistant.say("Hey, how can I help you today?", allow_interruptions=True)